fix code smells and duplication
